refactor: Rename package 'component' to 'components' (#43)
